requirements.txt
fastapi
uvicorn
pydantic>=2.0
openpyxl
google-generativeai
hypothesis
pytest
httpx
python-multipart
